Remove `Cluster` abstraction
