--- Website/ui/src/classes/clusters/Cluster.js
@@ -1,24 +1,0 @@
-import { City } from '@/services/CityService'
-
-export class Cluster {
-    constructor() {}
-
-    fromJson(jsonData) {
-        this.id = jsonData.id
-        this.name = jsonData.name
-        this.manager = jsonData.manager
-        if ('cities' in jsonData) {
-            this.cities = this.fetchCities()
-        }
-    }
-
-    fetchCities(cities) {
-        let result = []
-        for (let i in cities) {
-            let cityData = cities[i]
-            let city = new City()
-            city.fromJson(cityData)
-            result.push(city)
-        }
-    }
-}
